Revert "Merge pull request #7361 from hrydgard/color-conv-centralize"
This reverts commit f1b57dabf52f40d7f3ef6a6036279a2d8b473c7d, reversing
changes made to 41001637ce7af49cb5fbbd23b81e827eb1a18b3b.

+++ Common/Arm64Emitter.cpp
@@ -1512,7 +1512,7 @@ void ARM64XEmitter::ABI_PushRegisters(BitSet32 registers)
 		{
 			if (first)
 			{
-				STR(INDEX_PRE, (ARM64Reg)(X0 + it), SP, (u32)-(s32)stack_size);
+				STR(INDEX_PRE, (ARM64Reg)(X0 + it), SP, -stack_size);
 				first = false;
 				current_offset += 16;
 			}
